pushed old devices in a separate directory made compilation such as will use only wxWidgets for plots unless told differently Most probably will not compile on some platforms
