+++ src/otherdevices/devicewin.hpp
