+++ src/otherdevices/gdlwinstream.hpp
